requirements.txt
pandas
streamlit
openpyxl
matplotlib
reportlab
scikit-learn
numpy
plotly
seaborn
xlsxwriter
pdfplumber
python-docx
tabulate
arabic_reshaper
python-bidi
